getContentLength should return long

+++ org.osgi.framework/src/org/osgi/framework/connect/ConnectContent.java
@@ -118,7 +118,7 @@ public interface ConnectContent {
 		 * 
 		 * @return the size of the entry
 		 */
-		public int getContentLength();
+		public long getContentLength();
 
 		/**
 		 * Returns the last modification time of the entry
